WIP: this builds and runs, but seems very broken
I've not been able to successfully connect to a remote libvirt
instance, and adding a new local VM hangs with libosinfo not being
given the path to the .iso.

Obviously to get any further with running local virtualisation
instance, kvm needs to be added to the build and libvirtd needs
to be started up in some kind of user mode.

No cleanups during/after the build so it's pretty rough, and
Boxes definitely needs some more thought about how to present
useful functionality / graceful fallbacks when GUdev is not
available.